+++ CHANGELOG.md
@@ -6,4 +6,7 @@
 * add a node to deserialize a recipe to nbt
 * add an option for geckolib renderer to automatically `Schedule State Animation` without the node graph.
 * add multiblock debugger mod to tell why a multiblock is not assembling
+* machines formed with a catalyst candidate, will back to the original block when it is unformed
+* add a drop machine item option (when broken)
+  - machines formed with a catalyst candidate will drop the original block now